fieldset = fieldset_phy + field_NPP

+++ src/TurtleClass.py
@@ -13,12 +13,12 @@ def define_Turtle_Class(fieldset, param):
     class turtle(JITParticle):
         #Sampling
         if param['key_alltracers']:
-            T = Variable('T', to_write=True, dtype=np.float64)
-            NPP = Variable('NPP', to_write=True, dtype=np.float64)
+            T = Variable('T', to_write=True, dtype=np.float32)
+            NPP = Variable('NPP', to_write=True, dtype=np.float32)
         #All particles
         active = Variable('active', to_write=True, initial=1, dtype=np.float32)
-        prev_lon = Variable('prev_lon', to_write=False, dtype=np.float32, initial=attrgetter('lon'))
-        prev_lat = Variable('prev_lat', to_write=False, dtype=np.float32, initial=attrgetter('lat'))
+        prev_lon = Variable('prev_lon', to_write=False, dtype=np.float64, initial=attrgetter('lon'))
+        prev_lat = Variable('prev_lat', to_write=False, dtype=np.float64, initial=attrgetter('lat'))
         u_current = Variable('u_current', to_write=True, dtype=np.float32)
         v_current = Variable('v_current', to_write=True, dtype=np.float32)
         age = Variable('age', to_write=True, dtype=np.float64, initial=0.)
@@ -32,22 +32,22 @@ def define_Turtle_Class(fieldset, param):
             lethargy_time = Variable('lethargy_time', to_write=False, dtype=np.float32, initial=0.) #time spent under Tmin
             cold_death = Variable('cold_death', to_write=True, dtype=np.float32, initial=0)
         if param['cold_death'] or param['mode'] == 'active':
-            Tmin = Variable('Tmin', to_write=False, dtype=np.float64, initial=fieldset.Tmin)
+            Tmin = Variable('Tmin', to_write=False, dtype=np.float32, initial=fieldset.Tmin)
         #Active
         if param['mode'] == 'active':
-            u_swim = Variable('u_swim', to_write=True, dtype=np.float64)
-            v_swim = Variable('v_swim', to_write=True, dtype=np.float64)
-            M = Variable('M', to_write=False, dtype=np.float64)
-            vmax = Variable('vmax', to_write=False, dtype=np.float64)
-            PPmax = Variable('PPmax', to_write=True, dtype=np.float64)       
-            habT = Variable('habT', to_write=True, dtype=np.float64)
-            habPP = Variable('habPP', to_write=True, dtype=np.float64)
-            hab = Variable('hab', to_write=True, dtype=np.float64)
-            theta = Variable('theta', to_write=False, dtype=np.float64)
-            xgradh = Variable('xgradh', to_write=True, dtype=np.float64)
-            ygradh = Variable('ygradh', to_write=True, dtype=np.float64)
-            SCL = Variable('SCL', to_write=True, dtype=np.float64, initial=fieldset.SCL0)
-            Topt = Variable('Topt', to_write=False, dtype=np.float64, initial=fieldset.Topt)
+            u_swim = Variable('u_swim', to_write=True, dtype=np.float32)
+            v_swim = Variable('v_swim', to_write=True, dtype=np.float32)
+            M = Variable('M', to_write=False, dtype=np.float32)
+            vmax = Variable('vmax', to_write=False, dtype=np.float32)
+            PPmax = Variable('PPmax', to_write=True, dtype=np.float32)       
+            habT = Variable('habT', to_write=True, dtype=np.float32)
+            habPP = Variable('habPP', to_write=True, dtype=np.float32)
+            hab = Variable('hab', to_write=True, dtype=np.float32)
+            theta = Variable('theta', to_write=False, dtype=np.float32)
+            xgradh = Variable('xgradh', to_write=True, dtype=np.float32)
+            ygradh = Variable('ygradh', to_write=True, dtype=np.float32)
+            SCL = Variable('SCL', to_write=True, dtype=np.float32, initial=fieldset.SCL0)
+            Topt = Variable('Topt', to_write=False, dtype=np.float32, initial=fieldset.Topt)
             if param['growth'] == 'Gompertz':
                 K = Variable('K', to_write=False, dtype=np.float32, initial=fieldset.K0)
